Inline BasisFunctionSet Quadrature functions
These functions only passed the this pointer of the BasisFunctionSet to
the quadrature rule. By inlining these there are no more references to
QuadratureRule in BasisFunctionSet

 - Expansion was removed as it was deprecated and used the old
   functionality of BasisFunctionSet.
 - The inlining of the Curl requires some specialization in
   GaussQuadratureRule to handle the 2D/3D difference
 - Importing GaussQuadratureRules into BasisFunctionSet.h pulled in some
   of the requirements two filesi, these needed to be made explicit.
