+++ tests/unit/Utilities/020Nedelect2D_UnitTest.cpp
@@ -38,6 +38,7 @@
 
 #include "Utilities/BasisFunctions2DNedelec.h"
 #include "Integration/QuadratureRules/AllGaussQuadratureRules.h"
+#include "Integration/QuadratureRules/GaussQuadratureRule.h"
 #include "Geometry/ReferenceTriangle.h"
 #include "Base/BasisFunctionSet.h"
 #include "../catch.hpp"
